+++ component/prometheus/scrape/scrape.go
@@ -12,8 +12,8 @@ import (
 	component_config "github.com/grafana/agent/component/common/config"
 	"github.com/grafana/agent/component/discovery"
 	"github.com/grafana/agent/component/prometheus"
+	"github.com/grafana/agent/internal/useragent"
 	"github.com/grafana/agent/pkg/flow/logging/level"
-	"github.com/grafana/agent/pkg/useragent"
 	"github.com/grafana/agent/service/cluster"
 	"github.com/grafana/agent/service/http"
 	"github.com/grafana/agent/service/labelstore"
@@ -27,7 +27,7 @@ import (
 )
 
 func init() {
-	scrape.UserAgent = useragent.UserAgent()
+	scrape.UserAgent = useragent.Get()
 
 	component.Register(component.Registration{
 		Name:          "prometheus.scrape",
