+++ component/pyroscope/scrape/scrape_loop.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/grafana/agent/component/pyroscope"
+	"github.com/grafana/agent/internal/useragent"
 	"github.com/grafana/agent/pkg/flow/logging/level"
-	"github.com/grafana/agent/pkg/useragent"
 	commonconfig "github.com/prometheus/common/config"
 	"github.com/prometheus/prometheus/discovery/targetgroup"
 	"github.com/prometheus/prometheus/util/pool"
@@ -22,7 +22,7 @@ import (
 
 var (
 	payloadBuffers  = pool.New(1e3, 1e6, 3, func(sz int) interface{} { return make([]byte, 0, sz) })
-	userAgentHeader = useragent.UserAgent()
+	userAgentHeader = useragent.Get()
 )
 
 type scrapePool struct {
