+++ component/remote/http/http.go
@@ -13,13 +13,13 @@ import (
 	"github.com/go-kit/log"
 	"github.com/grafana/agent/component"
 	common_config "github.com/grafana/agent/component/common/config"
+	"github.com/grafana/agent/internal/useragent"
 	"github.com/grafana/agent/pkg/flow/logging/level"
-	"github.com/grafana/agent/pkg/useragent"
 	"github.com/grafana/river/rivertypes"
 	prom_config "github.com/prometheus/common/config"
 )
 
-var userAgent = useragent.UserAgent()
+var userAgent = useragent.Get()
 
 func init() {
 	component.Register(component.Registration{
